Move Schedulers from Scheduler namespace
