+++ Source/EasyNetQ.IntegrationTests/Utils/BusExtensions.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using EasyNetQ.Scheduling;
 
 namespace EasyNetQ.IntegrationTests.Utils
 {
