+++ Source/EasyNetQ.Tests/Scheduling/SchedulingExtensionsTests.cs
@@ -1,7 +1,6 @@
 // ReSharper disable InconsistentNaming
 
 using EasyNetQ.DI;
-using EasyNetQ.Scheduling;
 using NSubstitute;
 using Xunit;
